% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadJsonFiles2BQ.r
\name{loadJsonFiles2BigQuery}
\alias{loadJsonFiles2BigQuery}
\title{Load JSON files into a big query table}
\usage{
loadJsonFiles2BigQuery(project, dataset, table, source_uris,
  create_disposition = "CREATE_IF_NEEDED",
  write_disposition = "WRITE_APPEND")
}
\arguments{
\item{project}{Google cloud billing project}

\item{dataset}{Dataset that contains the table}

\item{table}{Table to load the json files into.}

\item{source_uris}{Vector containing the GS location of the files to be
loaded}
}
\description{
Load JSONL files that are stored in google cloud storage into
a big query table. Schema will be automatically detected.
}
\seealso{
\url{https://github.com/cloudyr/googleCloudStorageR} Files to be
     loaded can be uploaded easily using the cloudyr package

\url{https://cloud.google.com/bigquery/docs/reference/rest/v2/jobs#configuration.load}
}
